Respond with 200 OK and empty user for /api/user.
Do that instead of 401 Unauthorized and nil body.

This solves problem with 401 being redirected to /login (which doesn't
make sense for API calls).

It also seems to be friendlier and more consistent, since the 401
response shows up as an error in dev console. It's not an error,
there's just no authenticated user.