Refactor to use context from http.Request.
Start doing getUser(req) once at the beginning of a request handler,
then placing the *user in the request context. The Users.GetAuthenticatedSpec
method simply extracts the *user value from the context. This ensures
that the authenticated user is consistent for the duration of a request.
Otherwise it might be possible for a session to expire through midway
through the request handling. It should also more performant since
session map mutex is locked at exactly once for each request, rather
than potentially multiple times. This decision can be adjusted later
if desired.

It is an internal error if Users.GetAuthenticatedSpec is called on a
context where *user wasn't set. That can only happen if that initial
getUser(req) call and placing result in request context was forgotten
for a given handler.

A potential future enhancement is to factor out the initial
"call getUser(req) and place *user in the request context" step
into a user middleware.

Add that functionality to errorHandler and use it for more handlers.

Validate method for more routes.

Move towards changing users.Service.GetAuthenticatedSpec signature to
not return error, since it's unhelpful to return any error there to
callers (there can only be an internal error, which shouldn't happen;
dealing with it internally makes the most sense).