Treat error in authentication process as fatal for request.
Since the authentication is performed locally, an error can only occur
if there was an internal programming error. Still, it's better to
surface that in a standard way and not proceed with handling the
request as if no user was authenticated.

If authentication server is remote, then it's possible for
GetAuthenticatedSpec to return a non-nil error due to network
communication failure.