Remove unused mock test data.
This was primarily used and needed during early development stages of
the new index page with activity. Real data is used now, and should be
sufficient even for development (as long as one is online).

If this is needed in the future during development, it can be brought
back in a way such that it's actually used.

Followup to #4.