Display location of comments in activity stream.
Previously, comment events would only include the repository. Now, it
also displays the issue, PR, or commit that the comment is located in.

This provides more context to the comment and makes things more
readable. It also provides a link to the real comment for full details
and context.

Fetch additional commits from github.CommitCommentEvent events.

Factor out details aspect from various components to higher level. This
way, they can be used for details and for inline spans.