component: Remove tabindex="-1" from Header.
It was added without specific reason in initial implementation in 36f20db7806ffeb50a599e99f66dc477776b8882.
I don't see a good reason for it to exist. The link is still clickable
with mouse. Perhaps the link itself should be removed instead.

For now, just remove tabindex="-1" because it makes things a little
simpler.