Create AllowMethods helper; use http.Method* consts.
Trying this approach to learn if it's a good idea. The motivation
is to remove chance of mismatch between the method check, and the
returned method allowed slice. Previously, it was specified in two
places, but now only in one place.

Not sure if using http.Method* consts makes readability significantly
worse, or if it's acceptable. Will see over time.