Improve about page, make it mobile friendly.
Use experimental vec package for HTML rendering.

Make it mobile friendly by changing layout, so the photo is centered on
top, and the text follows underneath. Previous side-by-side layout
didn't scale well on its own.

Use a square avatar, put name and nickname under it. This is a pretty
common and more recognizable way to display a user.

Change description to be in 3rd person. Not sure if it's better in the
long term, so let's try it and see.

Change links to GitHub and Twitter to be more pleasant and readable.

Update Font Awesome to latest version 4.7.0.

Remove unused CSS; regenerate.