Make iconLink icon color an optional override.
Previously, by default, the icon color was black (i.e., RGB{}) and
would always override the current color. Make it an optional override.
If left with default value of nil, the icon color is inherited from
parent.

This makes the component more general, which will be used in next
commit.

Rename Color to IconColor to more accurately describe the field. It's
the color of the icon.