Add support for git push, auth via basic auth.
For now, the only accepted password is the same base64-encoded access
token as used to access the site via cookie. The username must match
the user login.

In the future, a more user friendly UI for creating personal access
tokens should be created.

At this time, pushing to kebabcase repo is only allowed for site admins
(i.e., shurcooL). Support for additional ways for other people to
contribute is a part of future work.