Change commit{,ID} components to expose entire hash.
Make only the first 8 characters visible, as before, but include the
entire hash in reality. This way, selecting it and copying it puts the
entire hash into the user's clipboard.

There should be no visible change aside from that behavior.

This is possible via CSS because we're using a monospace font (Go Mono).