Add commits tab for hosted Go package with vanity import path.
This allows seeing the history of the package and how it has changed,
as well as the rationale for changes in commit details. There are links
to viewing the code at each revision.

A future enhancement is to add a page for seeing the content of any
individual commit in full detail.

Regenerate ./assets.