Add invisible rel="me" link to GitHub profile on index page.
This is needed as a simple solution to be able to sign in to
https://indieweb.org. The link is not visible because there are no
planned visible links to my profile on the index page, those are
already located on /about page.

In the future, I can consider setting up my domain to support the
IndieAuth protocol, as described at https://indieweb.org/authorization-endpoint.

Updates #4.