internal/code: Use pointers to code.Directory rather than values.
This way, both the slice and map of views of directories refer to the
same structs. That makes it easier to modify them after creation (and
not have them go out of sync). That will be done in next commit.

Use a 2-element array of []*code.Directory parts in expandPattern,
rather than a 2-element slice. It should be ever so slightly simpler.