internal/code: Detect and track LICENSE files.
This change adds code to detect the license of each Go package.
It's based on LICENSE files found in directories of git repositories.
Whenever a LICENSE file is found, it applies to that directory
and all its child directories, unless a child directory has its
own LICENSE file that overrides the license further.

Add test for this behavior.