Display license for all hosted Go packages.
This change makes use of the work to detect the license of a Go package
in previous commit, and adds UI to display it. It appears as a "License"
link at the bottom of the package index page.

If a directory (nor any of its parent directories) doesn't provide a
specific LICENSE file, a default license located at root directory of
the domain is applied. That means if the user wishes to use the default
license for a new package or repository, they don't need to do any extra
work.

Use a new $file route to display the LICENSE file. At this time, no
other files can be displayed with it (that can be considered in the
future).