Display number of packages, open issues, open changes in repository tabnav.
Factor out the common code for creating a repository tabnav component
into repositoryTabnav function.

Create change service in main rather than initChanges. It needs to be
provided to other components as well now.

Counting issues in local issue trackers is not very fast because there's
no index yet. It's reasonably fast for now, but an index will need to be
added in the future as the number of issues grows. This change motivates
that work.