Move in code for rendering resume on frontend.
It was previously located in https://github.com/shurcooL/resume
repository, but that's not an optimal location for it.

The resume frontend rendering code needs some of home components
(e.g., the header) to render the entire page. Having its code be
in home repository is better, since it keeps related code closer.

Now, it's very consistent with other similar pages like /idiomatic-go
and /blog.