httputil: Don't gzip for status codes that can't have body.
Using a gzip writer and closing it has some fixed overhead when
the body is empty. We want empty body to remain empty, so avoid
using gzip encoding for status codes that are known not to have
a body.

This fixes these errors when response code is 304 and a body is
not expected:

	2018/06/15 21:46:46 warning: Content-Type header not set for POST "/blog/new"
	2018/06/15 21:46:46 gzipResponseWriter.Close: error closing *gzip.Writer: http: request method or response status code does not allow body

	2018/06/15 21:47:07 warning: Content-Type header not set for POST "/blog/26/edit"
	2018/06/15 21:55:59 gzipResponseWriter.Close: error closing *gzip.Writer: http: request method or response status code does not allow body

The problem could be reproduced with:

	curl -i -H 'If-Modified-Since: Sat, 16 Jun 2018 03:44:26 GMT' 'http://localhost:8080/usercontent/1924134@github.com/motion-blur-demo.png'