Create a new-repo action.
This action codifies and automates all the steps in the previously
manual process of creating a new repository. It is available only to
site admins.

It may not stay around unmodified for long, as there are plans to
consider making repository creation happen implicitly on first git push.
However, much of the code here will be refactored and reused for that.

A part of this change was created during the October 10th, 2018
Homebrew Website Club Meetup in NYC¹.

¹ https://indieweb.org/events/2018-10-10-homebrew-website-club#New_York_City