internal/code: Rediscover only within the affected repository.
This change is an optimization for code.Service.Rediscover method.
Instead of always rediscovering the entire store, limit scope to
just the affected repository.

Time to Rediscover with sample test data went from 13 ms to 6 ms.

This optimization will have a significant impact when the entire
repository store is very large compared to the single repository
being updated.

Increase test coverage, since some of the logic is non-trivial.