handle empty commits, show a blank slate
Previously, there was a panic when trying to display an empty commit,
because b[1:] was trying to trim a "\n" prefix that wasn't there.

Instead of stripping it with bytes.TrimPrefix(b, []byte{'\n'}), just
leave it. diff.ParseMultiFileDiff ignores it anyway.

Also factor out blank slate component, since it's used in many places.