add support for viewing history of a single Go package
It is implemented by specifying a git pathspec that constraints
git operations to only the directory that contains the target
Go package, and filters everything else out. Specifically:

	# Commits affecting files in root directory only.
	git log -- ':(glob)*'

	# Commits affecting files in dir subdirectory only.
	git log -- ':(glob)dir/*'

Currently, there is no UI that leads to the new pages. That will
be addressed in the future, when there are more package-specific
operations supported. This is a single step towards providing a
complete experience for individual packages.

Reference: https://git-scm.com/docs/gitglossary.html#def_pathspec.