replace generic -production flag with specific flags
It's better to be explicit and clear with the flags. Instead of having
one catch-all -production flag that controls many things, replace it
with three specific, targeted flags. The new flags include -store-dir,
-analytics-file, and -secure-cookie.

Also rename -statefile flag to -state-file for consistency.

This change updates home to follow more modern best practices,
and makes it consistent with other, newer codebases.