change module proxy URL to contain a non-empty host
Previously, the URL of the module proxy used to serve modules beginning
with the path dmitri.shuralyov.com contained an empty host value.
This worked successfully in Go 1.11 and Go 1.12, leading to GET requests
without stuttering. For example, to fetch a list of versions in the
dmitri.shuralyov.com/state module, the request was:

	GET https://dmitri.shuralyov.com/state/@v/list

However, a module proxy URL with an empty host no longer works in Go tip
as of https://golang.org/cl/170879, and evidence suggests that is the
better behavior to continue to enforce.

There are plans to document the requirement of a module proxy URL to
have a non-empty host. It's also planned to improve the error reporting
when the go command encounters an invalid module proxy URLs, so users
will have a better idea of what went wrong. See issue golang/go#32006.

To be compatible with Go 1.11, 1.12, and the future Go 1.13, this commit
changes the module proxy URL used to serve own modules from "https://"
to "https://dmitri.shuralyov.com/api/module". Future module proxy
requests will look like:

	GET https://dmitri.shuralyov.com/api/module/dmitri.shuralyov.com/state/@v/list

The module proxy protocol is meant primarily for machines rather than
humans, so the longer path that stutters is an acceptable compromise.

A benefit of the new proxy URL is that we no longer need to check
if a request matches. Instead, all HTTP requests with a "/api/module/"
path prefix can be immediately recognized as requests to the module
proxy, and handled as such.

As part of embracing the machine-focused aspect of these URLs,
also remove the optional human-friendly behavior of redirecting
when the module path and version are not encoded in the request URL.
Instead, serve a 400 Bad Request error. A correct module proxy client
will not make the mistake of forgetting to encode those values.

Updates shurcooL/home#24
Updates golang/go#32006