all: improve httptest usage
In tests, use the Result method. It produces output more representative
of what the handler actually responded with, making tests more accurate.

In legacy non-test usage as a temporary buffer, use Header() method
instead of relying on directly setting the deprecated HeaderMap map.
It was deprecated primarily for test-focused usage, but it's still
deprecated, and makes tools report using it as a problem.

Updates golang/go#32136
Updates golang/go#32144