add IssueEventURL and PullRequestEventURL methods
These methods have been added to github.Router interface.
Add them to dmitshurSeesHomeRouter and homeRouter so that
the interface continues to be implemented.

The "#event-%d" anchor in the home router is not implemented yet,
but start using it eagerly. It will work once this is implemented
in the local issues and changes apps. Only Dmitri will see these
links, so it's not going to be confusing to anyone else.

Follows https://dmitri.shuralyov.com/route/github$commit/7e79d56078638d1d7d4d6853af037d67de3f9589.