internal/exp/service: begin the notification service v2 initiative
This is an evolution of the notification service. Compared to v1,
the key difference is that individual notifications are tracked,
and there is support for real-time streaming of notifications.

Define a new activity service. It combines notification service v2
and events service v1 into one service. It turns out to be often
easier and more efficient to implement both at once. There is an
opportunity to explore unifying common aspects of notifications
and events further in the future.

Add implementations for GitHub and Gerrit. They are powered primarily
by notification emails sent by those services. They use relevant API
clients or services to augment information missing from notification
emails. The implementations are heavily tested and iterated on, but
they are still rough in some parts and need to be improved. That will
be done gradually over time, balanced with other priorities.

A local implementation, based on a virtual filesystem or some other
storage mechanism, still needs to be done.

Implement a v2-to-v1 adapter as well. It will be used to provide
Gerrit notifications to notifications app v1.

Implement HTTP client, handler, and route packages.