internal/exp/app: add initial notifications app v2 implementation
This new app is based on github.com/shurcooL/notificationsapp,
but uses notification service v2 and implements a new stream view.
The stream view displays individual notifications in real-time as
they arrive, and is the default tab. The traditional thread view
is available on another tab.

WebAssembly is used to render the app on the frontend.

The UI and API are still rough around the edges and will continue
to be iterated on over time, balanced with other priorities.