improve display of activity and notification streams
The overview is made to always fit into a single line, so that long
titles are cut off with an ellipsis rather than overflowing.

The two streams are made more consistent (and in the future can be
combined into one common rendering code).

Fixes https://github.com/shurcooL/home/issues/33.