add an IndieAuth authorization endpoint
This change implements an IndieAuth¹ server and adds it to home.
It enables Dmitri Shuralyov to sign in to other websites that
support the IndieAuth protocol via my own authorization endpoint.

This can work in a situation where github.com is down or otherwise
unavailable. It also gives me more control over whether to proceed
with signing in to some website or not (e.g., for now, the consent
screen is always shown on each sign in, rather than just the first
time).

¹ https://indieauth.spec.indieweb.org/

Updates https://github.com/shurcooL/home/issues/34.