preserve return parameter during sign in
This change fixes the behavior of returning to
the page the visitor was on after signing in.

An earlier change to move to URL-based sign in flow,
in commit ebebcaa7af0b3ca608f7b431ab797db35828c790,
added the action="/login" parameter to the sign in form.

Since the return parameter was being set inside the URL query,
that change caused the return parameter to not be included when
the "Sign In" button was pressed and the form was submitted.

Populate the return parameter in the query of the form action URL,
so that it will resume being included in the POST request.

Also start using req.RequestURI instead of req.URL.String() as the
value for returnParameterName in two more places, for consistency.

Updates https://github.com/shurcooL/home/issues/34.