add support for signing in with custom URLs via IndieAuth or RelMeAuth
This change implements and starts using an IndieAuth¹ client. With it,
users can now sign in with any custom URL that they own (typically
their personal website), as long as it supports the IndieAuth protocol.

This will work in situations where github.com is down or otherwise
unavailable. It also makes it possible for a user who does not have
a GitHub account to sign in without having to create such an account.

For convenience, there is also support for signing in with custom
URLs based on the RelMeAuth² protocol, which can be implemented just
by adding a <link href="https://github.com/example" rel="me">
link to the HTML page, and ensuring the the GitHub profile links
back to the same page to establish a bidirectional connection.

RelMeAuth allows signing in with a custom URL while delegating the
work of performing authentication to an existing OAuth 2.0 endpoint.
There is support only for the GitHub provider, with no plans to add
more at this time.

A fetch service is added and used to fetch information about users,
both from the independent web and from GitHub. There are two service
implementations: a direct one, and another that delegates the work
of doing fetches to a remote Google Cloud Function³.

There is a unified flow for all of IndieAuth, RelMeAuth, and GitHub
authentication:

1.	The first step is to fetch information about the user based on
	their user profile URL, and to look for valid bidirectionally
	connected profiles.

2.	Second step is to figure out what authentication methods are
	available for that user, and pick one.

3.	Third step is to perform the authentication process and verify
	that the authenticated user equals to the originally entered one.

This unified approach leads to fewer distinct code paths, and makes
handling and reporting of errors easier and safer to implement.

¹ https://indieauth.spec.indieweb.org/
² http://microformats.org/wiki/relmeauth
³ https://cloud.google.com/functions

Fixes https://github.com/shurcooL/home/issues/34.