move IndieAuth authorization endpoint advertisement to HTTP header
This should be more efficient and more secure compared to advertising
it in the HTML body.

A compliant IndieAuth client needs to check HTTP Link headers first,
HTML <link> elements second.¹ Advertising the endpoint in the header
means it will be done faster, and parsing HTML is not necessary.
An HTTP Link header can also be fetched as part of a HEAD request,
while HTML <link> elements require a GET request.

In the unlikely case of unexpected HTML content being injected into
the index page, either by mistake or malice, having the authorization
endpoint be advertised in the HTTP headers makes its value more
resistant to being overridden.

Also add a test case to catch regressions.

Thanks to @FiloSottile for the initial idea to consider this.

¹ https://indieauth.spec.indieweb.org/#discovery-by-clients

Updates https://github.com/shurcooL/home/issues/34.