internal/code: reject pseduo-versions from non-master branches
The intention of the current ModuleHandler is to serve only v0.0.0
pseudo-versions from commits that are on the master branch. This
requires checking the commit hash component of the incoming module
request, and verifying that the corresponding commit is a part of
the master branch.

This wasn't done previously, which means an explicit request to a
pseudo-version based on a commit on non-master branch would be served
by ModuleHandler as if such a version existed on the master branch.

No non-master branches were ever pushed, and pseudo-versions based
on commits on non-master branches were never advertised by the list
endpoint. As a result, unintentional versions were not indexed by the
Go module mirror.

Updates https://github.com/golang/go/issues/24031.