internal/exp/...issue...: move ListTimeline into Service, delete List{Comments,Events}
The ListTimeline method has proven to be a good idea by now.

Delete old ListComments and ListEvents methods and make the
ListTimeline method a non-optional part of Service interface
to simplify code.

Delete the issues.CopierFrom interface and its implementation
in the issue/fs package, because it requires maintenance but
is rarely used. It can be re-added if it's needed in the future.