internal/exp/cmd/spa: preserve URL fragment when navigating
As one example of why this change is helpful, consider clicking on
a notification pointing to a specific comment (which is a URL with
a fragment). Now that the URL fragment is preserved, it highlights
the target comment upon navigation, instead of just staying at the
top of the page.