support HEAD method in ?go-get=1 requests
The go command makes ?go-get=1 requests using only the GET method,
never with the HEAD method. HEAD requests to ?go-get=1 don't seem
very useful, since relevant information is placed in the response
body rather than headers, and it wouldn't be available.

However, the HTTP specification requires¹ that the HEAD method is
supported, so do so.

¹ https://tools.ietf.org/html/rfc7231#section-4