add /... suffix to repository home URL
It's better for the home page of a repository to be a page that displays
all packages contained by that repository, rather than a page that shows
the single package whose import path corresponds to the repository root.

Make that change by using an import path pattern with the "/..." suffix
as the home page URL.

Updates https://github.com/golang/go/issues/40477.