use wasm_exec.js file from GOROOT
Previously, the wasm_exec.js file had to be manually modified whenever
compiling with different major versions of Go. This change makes it so
that the wasm_exec.js file is used from the GOROOT directory of the go
command that is used during the build.

This way, no code changes and manual work needs to be done to support
rebuilding with newer Go versions. This is helpful to be able to test
betas and RCs of future Go versions sooner and more effortlessly.