internal/exp/app/notifsapp: make mark-read button keyboard accessible
Previously, the mark-read button was implement with an anchor HTML
element with an onclick handler. This type of button implementation
isn't possible to press with a keyboard, only with a mouse.

Reimplement the button using a button HTML element with an onclick
handler instead, so it's also accessible by keyboard.