internal/exp/service/notification/httpclient: retry in StreamNotifications
If there's a connection problem, add automatic retry with some backoff.
There's no reason to be silently failing and never retrying given that
StreamNotifications is expected to work until the provided context is
canceled.

This helps deal with dropped connection due to a temporary loss of
internet connectivity, or if the computer wakes up after sleeping.
It reduces the need for the user to manually refresh the browser page
to ensure that notifications are still streaming.